Huge modifications
Huge modifications compared to the first and second versions, as the program was divided into several classes, and many new features were added that were not present in previous versions, such as a special section for hotel management, this feature allowed many other features, such as controlling offers and prices by the hotel management, among the modifications. The other is an amendment to the method of storing data, as it has become more accurate and faster than previous versions.